CPM-main/requirements.txt
albumentations==0.5.2
blend-modes==2.1.0
dlib
dominate==2.6.0
numpy
opencv-python==4.5.1.48
Pillow
pretrainedmodels==0.7.4
scikit-image
segmentation-models-pytorch==0.1.3
tensorboard==1.14.0
tensorboardX==2.2
tensorflow-gpu==1.14.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.59.0
